requirements.txt
flask>=2.3.3
python-dotenv>=1.0.0
google-generativeai>=0.7.0
gitpython>=3.1.40
gunicorn>=21.2.0
requests>=2.31.0